refactor: 스터디 편집 및 탭 섹션 컴포넌트 리팩토링 #112
- 스터디 편집 페이지 및 관련 폼/컴포넌트 제거, 강의/공지사항/멤버를 위한 모듈형 탭 컴포넌트로 탭 섹션 구현 교체
- 미사용 스터디 편집 엔드포인트 제거를 위한 API 레이어 업데이트
- 유지보수성 향상을 위해 다이얼로그 래퍼 및 탭 로직을 새로운 구조로 구성
