+++ apps/intra/src/features/study/pages/study-detail/desktop-study-detail-page.tsx
@@ -2,7 +2,7 @@
 
 import { BackButton, Button, LoadingDots, StudyInfoSection } from '@hiarc-platform/ui';
 import { useStudyDetailPageState } from '@/features/study/hooks/page/use-study-detail-page-state';
-import { TabSection } from '@/features/study/components/tab-section';
+import { TabSection } from '@/features/study/components/tab-section/TabSection';
 
 export function DesktopStudyDetailPage(): React.ReactElement {
   const {
