--- apps/intra/src/features/study/pages/study-edit/index.ts
@@ -1,2 +1,0 @@
-export { DesktopStudyEditPage } from './desktop-study-edit-page';
-export { MobileStudyEditPage } from './mobile-study-edit-page';
